pyproject.toml
[tool.poetry]
name = "smartbridge-comms"
version = "0.0.15"
description = "This is proxy that facilitates the comms between the InnfluxDB cloud and the SmartBridge engine"
authors = ["angelos <benderbendingrodriguezsr@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.6.7"
influxdb-client = "^1.10.0"
PyYAML = "^5.3.1"
pandas = "^1.1.2"
DateTime = "^4.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
mypy = "^0.782"
black = "^20.8b1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
